Moved DrawPoint files.
