+++ samples/unit-tests/series-treemap/draw-point/demo.js
